requirements.txt
google-pasta
grpcio
psutil
opencv-python
numpy
torch
torchvision
tensorflow
tensorflow-hub
ultralytics
matplotlib
requests
Pillow